update 23/4 12h

+++ src/main/java/com/example/BE_employees_performance/entity/Kpa.java
@@ -23,7 +23,7 @@ public class Kpa  {
 
     @Lob
     private String description;
-    private Integer percent;
+    private Float percent;
 
     @ManyToOne
     @JoinColumn(name = "kpi_id")
